requirements.txt
aioredis==1.3.1
amqp==2.6.1
asgiref==3.4.1
async-timeout==4.0.2
attrs==22.2.0
autobahn==23.1.2
Automat==22.10.0
azure-core==1.26.4
azure-identity==1.12.0
azure-storage-blob==12.16.0
billiard==3.6.4.0
bleach==3.1.5
CacheControl==0.12.6
cachetools==4.2.4
celery==4.4.7
certifi==2021.5.30
cffi==1.15.1
channels==2.4.0
channels-redis==3.2.0
charset-normalizer==2.0.6
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
crc16==0.1.1
crypto==1.4.1
cryptography==39.0.0
daphne==2.5.0
Deprecated==1.2.13
Django==3.2.6
django-celery-beat==2.1.0
django-common-helpers==0.9.2
django-cors-headers==3.10.0
django-cron==0.5.1
django-crontab==0.7.1
django-debug-toolbar==3.2.2
django-extensions==3.1.3
django-filter==2.4.0
django-guardian==2.3.0
django-redis==4.12.1
django-rest-swagger==2.2.0
django-timezone-field==4.2.3
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.2
drf-yasg==1.20.0
emoji==0.6.0
facebook-sdk==3.1.0
fcm-django==0.3.4
firebase-admin==5.0.3
flower==0.9.4
future==0.18.2
google-api-core==2.8.2
google-api-python-client==2.80.0
google-auth==2.16.2
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.5.0
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.56.4
grpcio==1.41.0
grpcio-status==1.41.0
hiredis==2.2.2
httplib2==0.19.1
humanize==0.5.1
hyperlink==21.0.0
idna==3.2
incremental==22.10.0
inflection==0.5.1
infobip-api-python-client==3.0.1
install==1.3.4
isodate==0.6.1
itypes==1.2.0
Jinja2==3.0.1
kombu==4.6.11
MarkupSafe==2.0.1
mixins==0.1.4
msal==1.22.0
msal-extensions==1.0.0
msgpack==1.0.2
mysqlclient==2.0.1
Naked==0.1.31
numpy==1.22.1
openapi-codec==1.3.2
packaging==21.0
parse==1.16.0
pendulum==2.1.2
Pillow==8.0.1
portalocker==2.7.0
proto-plus==1.19.2
protobuf==3.18.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.10.3
pyfcm==1.4.7
PyJWT==2.6.0
pymongo==3.11.2
PyMySQL==1.0.2
pyOpenSSL==23.0.0
pyparsing==2.4.7
python-crontab==2.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-memcached==1.59
pytz==2021.3
pytzdata==2020.1
PyYAML==6.0
qrcode==6.1
redis==4.0.2
requests==2.26.0
rsa==4.7.2
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
schedule==1.1.0
serializers==0.2.4
service-identity==21.1.0
shellescape==3.8.1
simplejson==3.17.5
six==1.16.0
sqlparse==0.4.2
status==0.2.5
stringcase==1.2.0
tornado==6.2
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.5.0
uritemplate==3.0.1
urllib3==1.26.7
vine==1.3.0
webencodings==0.5.1
whitenoise==5.3.0
wrapt==1.15.0
zope.interface==5.5.2
